Refactor : party member 패키지 도입
- domain 패키지 내에 partyMember 패키지를 만들었습니다.
- 기존에는 partyMember 엔티티와 repo는 빵판 패키지에 있고 service는 파티 패키지에 있어서 관리가 어려웠습니다.
- party member가 빵판에 있는 것도, 파티에 있는 것도 적절하지 않다고 판단하였습니다. 두 영역에 다 속하기 때문입니다.
- 따라서 패키지 하나를 새로 만들었습니다.
- 도메인은 사용자가 프로그램을 사용하는 대상 분야라고 합니다. 파티 참여자로서의 역할이 파티멤버 도메인이라고 생각하여 한 도메인으로 봐도 될 것 같습니다. .
