+++ src/main/java/nbbang/com/nbbang/domain/party/repository/PartyRepositorySupportImpl.java
@@ -2,14 +2,11 @@ package nbbang.com.nbbang.domain.party.repository;
 
 import com.querydsl.jpa.impl.JPAQueryFactory;
 import lombok.RequiredArgsConstructor;
-import nbbang.com.nbbang.domain.bbangpan.domain.QPartyMember;
 import nbbang.com.nbbang.domain.member.dto.Place;
 import nbbang.com.nbbang.domain.party.domain.Party;
 import nbbang.com.nbbang.domain.party.domain.PartyStatus;
-import nbbang.com.nbbang.domain.party.domain.QParty;
 
-import static nbbang.com.nbbang.domain.bbangpan.domain.QPartyMember.partyMember;
-import static nbbang.com.nbbang.domain.party.domain.QParty.party; // JPA 책 432p.
+import static nbbang.com.nbbang.domain.party.domain.QParty.party;
 
 import java.util.List;
