+++ src/main/java/nbbang/com/nbbang/domain/partyMember/dto/PartyMemberResponseDto.java
@@ -1,19 +1,8 @@
-package nbbang.com.nbbang.domain.bbangpan.dto;
+package nbbang.com.nbbang.domain.partyMember.dto;
 
 import lombok.Builder;
 import lombok.Data;
-import nbbang.com.nbbang.domain.bbangpan.domain.PartyMember;
-import nbbang.com.nbbang.domain.bbangpan.domain.SendStatus;
-import nbbang.com.nbbang.domain.chat.domain.Message;
-import nbbang.com.nbbang.domain.member.domain.Member;
-import nbbang.com.nbbang.domain.party.domain.Party;
-
-import javax.persistence.Enumerated;
-import javax.persistence.JoinColumn;
-import javax.persistence.ManyToOne;
-import javax.persistence.OneToOne;
-
-import static javax.persistence.EnumType.STRING;
+import nbbang.com.nbbang.domain.partyMember.domain.PartyMember;
 
 @Builder
 @Data
