+++ src/main/java/nbbang/com/nbbang/domain/partyMember/repository/PartyMemberRepository.java
@@ -1,8 +1,7 @@
-package nbbang.com.nbbang.domain.bbangpan.repository;
+package nbbang.com.nbbang.domain.partyMember.repository;
 
-import nbbang.com.nbbang.domain.bbangpan.domain.PartyMember;
+import nbbang.com.nbbang.domain.partyMember.domain.PartyMember;
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
 
 
 public interface PartyMemberRepository extends JpaRepository<PartyMember, Long>, PartyMemberRepositorySupport {
