+++ src/main/java/nbbang/com/nbbang/global/cache/PartyMemberIdCache.java
@@ -4,7 +4,7 @@ import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import nbbang.com.nbbang.domain.bbangpan.domain.PartyMember;
+import nbbang.com.nbbang.domain.partyMember.domain.PartyMember;
 
 @Getter
 @AllArgsConstructor
